mapeo de base de datos

+++ Cruz_FrancoMatias_2931_TP4/src/main/java/ar/edu/unju/fi/tp4/service/imp/ProductoServiceImp.java
@@ -63,7 +63,7 @@ public class ProductoServiceImp implements IProductoService{
 	}
 
 	@Override
-	public Producto getProductoporCodigo(int codigo) {
+	public Producto getProductoporCodigo(long codigo) {
 		Producto produ = new Producto();
 		for(Producto pro: productos) {
 			if(pro.getCodigo() == codigo) {
